requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.4.0f
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
pkg-resources==0.0.0
PyJWT==1.7.1
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
